+++ dbt/macros/b3.sql
@@ -1,0 +1,3 @@
+{% macro b3_fonte_oficial() %}
+  https://www.b3.com.br/pt_br/market-data-e-indices/servicos-de-dados/market-data/historico/boletins-diarios/pesquisa-por-pregao/
+{% endmacro %}