+++ dbt/models/marts/b3/mart_b3__futuro_boi.sql
@@ -1,0 +1,7 @@
+{{ config(
+    materialized='external',
+    location=var('gold_root') ~ '/mart_b3__futuro_boi/mart.parquet',
+    format='parquet'
+) }}
+
+select * from {{ ref('stg_b3__futuro_boi') }}